requirements.txt
Django==1.10
django-cors-headers==2.1.0
django-crispy-forms==1.6.0
django-filter==0.14.0
django-webpack-loader==0.4.1
djangorestframework==3.4.4
Markdown==2.6.6
psycopg2==2.6.2
